implemented comment deleting only for comment author and admins

+++ ASP.NET-MVC-Forum/ASP.NET-MVC-Forum/Areas/API/Controllers/CommentsController.cs
@@ -62,6 +62,11 @@
                 return this.BadRequest();
             }
 
+            if (comment.UserId != await userService.GetBaseUserIdAsync(this.User.Id()) && !this.User.IsAdmin())
+            {
+                return StatusCode(401,"Only the comment author or site administrator can delete this comment");
+            }
+
             await commentService.DeleteCommentAsync(comment);
 
             return Ok();
